Get joint nextjs/socket.io dev/prod server working
