+++ app/.gitignore
@@ -1,2 +1,3 @@
 .next
 data
+dist
