+++ app/server/index.ts
@@ -1,21 +1,26 @@
 import { z } from 'zod'
+import http from 'http'
 
 export const Options = z.object({
-  next: z.object({
-    hostname: z.string().default('0.0.0.0'),
-    port: z.number().default(3000),
-  }).optional(),
-  ws: z.object({
-    hostname: z.string().default('0.0.0.0'),
-    port: z.number().default(3005),
-  }).optional(),
+  hostname: z.string().default('0.0.0.0'),
+  port: z.number().default(3000),
+  plugins: z.array(z.enum(['ws', 'next'])).default(['ws', 'next']).transform(L => new Set(L)),
 })
 
 export default async function main(opts: z.infer<typeof Options>) {
-  if (opts.ws) {
-    await (await import('./plugins/ws')).default(opts.ws)
+  const server = http.createServer()
+  if (opts.plugins.has('ws')) {
+    await (await import('./plugins/ws')).default(server, opts)
   }
-  if (opts.next) {
-    await (await import('./plugins/next')).default(opts.next)
+  if (opts.plugins.has('next')) {
+    await (await import('./plugins/next')).default(server, opts)
   }
+  server.once('error', (err) => {
+    console.error(err)
+    process.exit(1)
+  })
+  await new Promise<void>((resolve, reject) => {
+    server.listen(opts.port, () => {resolve()})
+  })
+  console.log(`Server listening on http://${opts.hostname}:${opts.port}`);
 }
