+++ package.json
@@ -8,7 +8,7 @@
     "./components/**"
   ],
   "scripts": {
-    "build": "npx --workspace=app next build",
+    "build": "npm --workspace=app run build",
     "codegen:apis": "ts-node cli/sync-api-import.ts",
     "codegen:components": "ts-node cli/sync-component-import.ts",
     "codegen:cite": "ts-node cli/add-citation.ts",
@@ -17,14 +17,13 @@
     "codegen:sql": "mkdir -p db/migrations && ts-node cli/build-sql.ts > db/migrations/00_init.sql",
     "codegen:table": "ts-node cli/build-table.ts > metanodes.tsv",
     "codegen:analytics": "ts-node cli/dump-analytics.ts > analytics.tsv",
-    "dev": "NODE_ENV=development npx --workspace=app ts-node server/cli.ts",
+    "dev": "npm run --workspace=app dev",
     "start:worker": "ts-node cli/resolve-process-worker.ts",
     "start:run-wes-worker": "ts-node cli/run-wes-worker.ts",
-    "start:wes-worker": "NODE_ENV=production ts-node cli/wes-worker.ts",
-    "start:ssl": "concurrently \"npm run start -- -p 3001\" \"ssl-proxy --source 3000 --target 3001\"",
-    "start": "NODE_ENV=production npx --workspace=app ts-node server/cli.ts",
-    "start:ui": "NODE_ENV=production npx --workspace=app ts-node server/cli.ts '{\"next\": {}}'",
-    "start:ws": "NODE_ENV=production npx --workspace=app ts-node server/cli.ts '{\"ws\": {}}'",
+    "start:wes-worker": "NODE_ENV=production DATABASE_URL= ts-node cli/wes-worker.ts",
+    "start": "npm run --workspace=app start",
+    "start:ui": "npm run --workspace=app start:ui",
+    "start:ws": "npm run --workspace=app start:ws",
     "version": "ts-node cli/version-compose.ts && git add docker-compose.yaml && docker-compose build --parallel",
     "postversion": "docker-compose push"
   },
@@ -34,7 +33,6 @@
     "@blueprintjs/select": "^5.0.16",
     "@blueprintjs/table": "^5.0.16",
     "@mdi/js": "^7.3.67",
-    "@next/bundle-analyzer": "^13.4.19",
     "@socket.io/postgres-adapter": "^0.3.1",
     "@types/pluralize": "^0.0.32",
     "axios": "^1.6.1",
